+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # QOwnNotes Web Companion Changelog
 
+## 18.12.1
+- fixed the default port
+- restructured the context menu
+- added an context menu entry to reset the settings
+
 ## 18.12.0
 - added support to add selected text as new note in QOwnNotes 18.12.7 and higher
   (for QOwnNotes issue [#330](https://github.com/pbek/QOwnNotes/issues/330))